requirements.txt
grpcio
grpcio-tools
protobuf
numpy
torch
onnxruntime
